Fix that rulers aren't dislocated by margin change
The drawn points are not aware of the set margin, only
restricted by it during drawing. If the margin is changed,
the canvas is changed, and the points are drawn with their
old coordinates on the new canvas, dislocating them relative
to the bounding box.
When changing the margin, the points therefore need to be scaled
by the change in margin too keep their original positions in the
image.